Rename classes
* 'ListUtil' to 'RichList'
* 'ResolverHelper' to 'DispatcherResolver'
* 'PublisherImplicits' to 'RichPubSubService'
* 'PublishSubscribeService' to 'PubSubService'
