+++ src/main/scala/common/RichList.scala
@@ -1,6 +1,6 @@
 package common
 
-object ListUtil {
+object RichList {
   implicit class ListSplit[A](list: List[A]) {
     def cutOff: (A, List[A]) = (list.head, list.tail)
   }
