+++ src/main/scala/common/actors/InterceptorBefore.scala
@@ -1,7 +1,7 @@
 package common.actors
 
 import akka.actor.{Actor, ActorLogging}
-import common.ListUtil._
+import common.RichList._
 import common.actors.Dispatcher.{Failure, InterceptorBeforeMessage, Success}
 import core.graphql.UserContext
