+++ src/main/scala/modules/counter/guice/modules/PublisherModule.scala
@@ -1,11 +1,11 @@
 package modules.counter.guice.modules
 
-import core.services.publisher.{PublishSubscribeServiceImpl, PublishSubscribeService}
+import core.services.publisher.{PubSubServiceImpl, PubSubService}
 import modules.counter.models.Counter
 import net.codingwell.scalaguice.ScalaModule
 
 class PublisherModule extends ScalaModule {
   override def configure() {
-    bind[PublishSubscribeService[Counter]].to[PublishSubscribeServiceImpl[Counter]]
+    bind[PubSubService[Counter]].to[PubSubServiceImpl[Counter]]
   }
 }
